requirements.txt
langchain-core
langchain-openai
langchain-community
langchain-ollama
langserve
streamlit
dotenv
huggingface_hub 
transformers
fastapi
uvicorn 
sse_starlette